+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: have both ECS Cluster and service to be created by one module
 - feat: add an example that logs to s3 bucket
 - feat: test container insights in complete example
+- feat: option to create one kms key for encryption inside the module
 
 
 ## [1.0.7] - 2023-02-01