+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: have both ECS Cluster and service to be created by one module
 - feat: add an example that logs to s3 bucket
 - feat: test container insights in complete example
-- feat: option to use cmk to encrypt ec2 ebs volumes 
+- feat: option to use cmk to encrypt ec2 ebs volumes
 
 ## [1.1.0] - 2023-03-08
 - feat: option to create one kms key for encryption inside the module